DAB-17: update unit vs func tests
